feat(excel): 修复浮点单元格类型读取为null的问题
- 将 EasyExcel、FastExcel 和 FesodExcelHelper 的支持类型由 NUMBER 改为 STRING，确保浮点单元格能正确读取
- ReadCellData 类新增 originalNumberValue，支持解析和校正浮点数精度问题
- 优化 IP 解析逻辑，支持多代理头中 unknown 的回退处理及 IPv6 环回地址转换
- 幂等切面异常处理中增加清理异常捕获，避免覆盖原始异常
- 本地锁实现改用引用计数管理，避免缓存淘汰时锁被误释放，保障高并发下的互斥性
- 添加分布式锁和幂等操作被 Redis 事务或 pipeline 包裹时的 WARN 提示，提醒用户避免误用
- 限流令牌桶计算修复，防止高速率叠加长时间空闲时整数溢出导致永久卡死
- 异常处理返回文案兜底，避免 null 出现和潜在空指针异常
- Feign 调用日志中未配置 traceId 时跳过处理，避免链路追踪丢失

+++ velo-autoconfigure-common/src/main/java/io/github/luminion/velo/ratelimit/support/CaffeineRateLimitHandler.java
@@ -42,7 +42,8 @@ public class CaffeineRateLimitHandler implements RateLimitHandler {
 
             long now = System.nanoTime();
             long nanosSinceLastRefill = now - lastRefillNanos;
-            double newTokens = nanosSinceLastRefill * capacity / (double) intervalNanos;
+            // 先转 double 再乘，避免 long * long 中间结果在高速率+长空闲时静默溢出为负导致令牌永不补充
+            double newTokens = (double) nanosSinceLastRefill * capacity / (double) intervalNanos;
             if (newTokens > 0D) {
                 tokens = Math.min(capacity, tokens + newTokens);
                 lastRefillNanos = now;
